refactor: move timer functions to hook useTimer
