+++ frontend/src/App.tsx
@@ -5,36 +5,12 @@ import { useEffect, useRef, useState } from 'react'
 import GameInfoBar from '../components/GameInfoBar.tsx'
 import Minefield from '../components/Minefield.tsx'
 import { GameProvider } from '../contexts/GameContext.tsx'
+import { useTimer } from '../hooks/useTimer.tsx'
 
 function App() {
+  const { time, gameIsRunning, startTimer, stopTimer } = useTimer()
+
   const [exploded, setExploded] = useState(false)
-
-  const [time, setTime] = useState(0)
-  const [gameIsRunning, setGameIsRunning] = useState<boolean | null>(null)
-  const timerRef = useRef<NodeJS.Timeout | null>(null)
-
-  const startTimer = () => {
-    if (!gameIsRunning) {
-      setGameIsRunning(true)
-      timerRef.current = setInterval(() => {
-        setTime((prev) => prev + 1)
-      }, 1000)
-    }
-  }
-
-  const stopTimer = () => {
-    setGameIsRunning(false)
-    if (timerRef.current) {
-      clearInterval(timerRef.current)
-      timerRef.current = null
-    }
-  }
-
-  useEffect(() => {
-    return () => {
-      if (timerRef.current) clearInterval(timerRef.current)
-    }
-  }, [])
 
   const [mineCount, setMineCount] = useState(40)
   const updateMineCount = (value: number) => {
